Added ticker to hash & send blocks
Added a ticker that calls `pushToBlockChainAndUpdate` every 5 minutes.
In the future this will be used to hash the latest block, add its hash
to the blockchain, and add a new block to the notarychain. This will
only occur if a new block has been added to the blockchain since the
last check and if there are entries waiting to be notarized.

A mutex is used to ensure that `pushToBlockChainAndUpdate` and `post`
will not enter a race condition.